Alter Boopsie::Email so that authentication to send email is optional.
If the user and/or password are not specified in smtp.json prefs, then
Boopsie::Email will attempt to send the email without authentication.
This is useful if your computer is allowed to relay through your mail
server without authenticating, first.  In this case, there is no need
to authenticate.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>